handle goto links to external files; removed old-style cast

+++ src/controlscreen.cpp
@@ -698,7 +698,7 @@ void ControlScreen::receiveCloseSignal()
 void ControlScreen::keyPressEvent(QKeyEvent* event)
 {
     // Key codes are given as key + modifiers.
-    QMap<int, QList<int>>::iterator map_it = keymap->find(event->key() + event->modifiers());
+    QMap<int, QList<int>>::iterator map_it = keymap->find(event->key() + static_cast<int>(event->modifiers()));
     if (map_it==keymap->end())
         return;
     for (QList<int>::const_iterator action_it=map_it->cbegin(); action_it!=map_it->cend(); action_it++) {
